Move prefilter to its own module and struct
