+++ nidx/src/searcher/shard_search.rs
@@ -13,7 +13,6 @@
 // limitations under the License.
 //
 
-use std::collections::HashSet;
 use std::sync::Arc;
 
 use nidx_json::JsonSearcher;
@@ -27,7 +26,7 @@ use tracing::{Span, instrument};
 use uuid::Uuid;
 
 use crate::errors::{NidxError, NidxResult};
-use crate::searcher::query_planner::{GraphIndexQueries, IndexQueries, PrefilterRequest};
+use crate::searcher::query_planner::{GraphIndexQueries, IndexQueries};
 
 use super::index_cache::IndexCache;
 use super::query_planner;
@@ -82,7 +81,7 @@ async fn shard_search(
         None
     };
 
-    let text_search = if query_plan.prefilter.texts.is_some() || query_plan.index_queries.texts_request.is_some() {
+    let text_search = if query_plan.prefilter.text.is_some() || query_plan.index_queries.texts_request.is_some() {
         let Some(text_index) = indexes.text_index() else {
             return Err(NidxError::NotFound);
         };
@@ -170,38 +169,6 @@ fn apply_shard_id_to_response(response: &mut SearchResponse, shard_id: Uuid) {
             result.shard_id = shard_id_bytes.clone();
         }
     }
-}
-
-fn compute_prefilter(
-    mut plan_prefilter: PrefilterRequest,
-    json_searcher: Option<&JsonSearcher>,
-    text_searcher: Option<&TextSearcher>,
-) -> anyhow::Result<PrefilterResult> {
-    let mut text_prefilter_result: Option<anyhow::Result<PrefilterResult>> = None;
-    let mut json_prefilter_result: Option<anyhow::Result<HashSet<Uuid>>> = None;
-
-    std::thread::scope(|scope| {
-        if let Some(prefilter) = plan_prefilter.texts.take() {
-            let current = Span::current();
-            let result = &mut text_prefilter_result;
-            scope.spawn(move || *result = Some(current.in_scope(|| text_searcher.unwrap().prefilter(&prefilter))));
-        }
-        if let Some(request) = plan_prefilter.json.take()
-            && let Some(searcher) = json_searcher
-        {
-            let current = Span::current();
-            let result = &mut json_prefilter_result;
-            scope.spawn(move || *result = Some(current.in_scope(|| searcher.search(&request))));
-        }
-    });
-
-    let text_prefilter = text_prefilter_result.transpose()?.unwrap_or(PrefilterResult::All);
-    let combined = if let Some(uuids) = json_prefilter_result.transpose()? {
-        text_prefilter.combine(uuids, plan_prefilter.filter_operator)
-    } else {
-        text_prefilter
-    };
-    Ok(combined)
 }
 
 #[allow(clippy::too_many_arguments)]
@@ -292,7 +259,7 @@ fn blocking_search(
 ) -> anyhow::Result<SearchResponse> {
     let mut index_queries = query_plan.index_queries;
 
-    let prefilter = compute_prefilter(query_plan.prefilter, json_searcher, text_searcher)?;
+    let prefilter = query_plan.prefilter.run(text_searcher, json_searcher)?;
     index_queries.apply_prefilter(prefilter);
 
     run_index_searches(
